// src/types/SuiviSuperControle.ts
    export interface Suivisupercontrole {
        id_supercontrole: number;
        code_artc: string;
        rev_projet: string;
        nom_client: string;
        type_controle: string;
        doc_refirance: string;
        methode_controle: string;
        date_debut: Date;
        duree_estime: Date;
        tracibilite_cablage: string;
        tracibilite_carton: string;
        heurs_internedepensees: number;
        date_final: Date;
    }